+++ packages/core/src/components/DialogBox/Actions/Actions.test.tsx
@@ -8,6 +8,7 @@ describe('DialogBox action component', () => {
         ['center', 'center'],
         ['right', 'flex-end']
     ];
+
     test.each(testData)('should align actions properly when align-items props is %s', (alignItems, flexValue) => {
         const { container } = render(<Actions alignItems={alignItems}>Demo Actions</Actions>);
         expect(container.querySelector('#default-id-actions')).toHaveStyle(`justify-content: ${flexValue}`);