+++ packages/core/src/components/Pagination/Pagination.test.tsx
@@ -1,4 +1,4 @@
-import { cleanup, fireEvent, render } from '@test-utils';
+import { cleanup, fireEvent, render, screen } from '@test-utils';
 import { Pagination } from './Pagination';
 
 describe('Pagination component', () => {
@@ -28,45 +28,45 @@ describe('Pagination component', () => {
     });
 
     it('should call onClick handler with correct page number when any page link is clicked', () => {
-        const { getByRole, mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
-        fireEvent.click(getByRole('button', { name: '4' }));
+        const { mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
+        fireEvent.click(screen.getByRole('button', { name: '4' }));
         expect(mockOnPageClick).toBeCalledWith(4);
     });
 
     it('should call onClick handler with first page when first page is clicked', () => {
-        const { getByRole, mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
-        fireEvent.click(getByRole('button', { name: '1' }));
+        const { mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
+        fireEvent.click(screen.getByRole('button', { name: '1' }));
         expect(mockOnPageClick).toBeCalledWith(1);
     });
 
     it('should call onClick handler with last page when last page is clicked', () => {
-        const { getByRole, mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
-        fireEvent.click(getByRole('button', { name: '15' }));
+        const { mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
+        fireEvent.click(screen.getByRole('button', { name: '15' }));
         expect(mockOnPageClick).toBeCalledWith(15);
     });
 
     it('should call onClick handler with prev page when prev link is clicked', () => {
-        const { mockOnPageClick, getByTitle } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
-        fireEvent.click(getByTitle('prev'));
+        const { mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
+        fireEvent.click(screen.getByTitle('prev'));
         expect(mockOnPageClick).toBeCalledWith(2);
     });
 
     it('should call onClick handler with next page when next link is clicked', () => {
-        const { mockOnPageClick, getByTitle } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
-        fireEvent.click(getByTitle('next'));
+        const { mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
+        fireEvent.click(screen.getByTitle('next'));
         expect(mockOnPageClick).toBeCalledWith(4);
     });
 
     it('should call onClick handler with expected page when selected from ellipsis popover', () => {
-        const { mockOnPageClick, getByRole } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
-        fireEvent.click(getByRole('button', { name: '...' }));
-        fireEvent.click(getByRole('button', { name: '7' }));
+        const { mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
+        fireEvent.click(screen.getByRole('button', { name: '...' }));
+        fireEvent.click(screen.getByRole('button', { name: '7' }));
         expect(mockOnPageClick).toBeCalledWith(7);
     });
 
     it('should not call onClick handler when clicked on active page', () => {
-        const { mockOnPageClick, getByRole } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
-        fireEvent.click(getByRole('button', { name: '3' }));
+        const { mockOnPageClick } = renderer({ totalItems: 150, activePage: 3, itemsPerPage: 10 });
+        fireEvent.click(screen.getByRole('button', { name: '3' }));
         expect(mockOnPageClick).not.toBeCalled();
     });
 });